requirements.txt
anyio==3.3.4
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==4.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fastapi==0.70.0
flake8==4.0.1
Flask==2.0.2
fonttools==4.28.4
h11==0.12.0
hexdump==3.3
httptools==0.2.0
idna==3.3
ifcfg==0.22
ipykernel==6.6.0
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
more-itertools==5.0.0
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
pandas==1.3.5
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==2.4.7
pyshark==0.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-evtx==0.7.4
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests==2.26.0
scapy==2.4.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.2
wget==3.2
yara-python==4.1.3
zipp==1.0.0
